feat(ui): update layout and manage lease css
