+++ frontend/src/components/Layout.tsx
@@ -1,15 +1,20 @@
 import { useState } from "react";
 import SideNavbar from "./navbar/SideNavbar";
-import { Navigate, Outlet, useLocation } from "react-router-dom";
-import { Menu } from "lucide-react";
+import { Navigate, Outlet, useLocation, useNavigate } from "react-router-dom";
+import { ArrowLeft, Menu } from "lucide-react";
 import useGlobalContext from "../hooks/useGlobalContext";
 
 const MainLayout = ({ usersRole }: { usersRole: string }) => {
   const { user } = useGlobalContext();
 
   const location = useLocation();
+  const navigate = useNavigate();
 
   const [isSidebarVisible, setIsSidebarVisible] = useState(false);
+
+  const goBack = () => {
+    navigate(-1);
+  };
 
   const toggleSidebar = () => {
     setIsSidebarVisible(!isSidebarVisible);
@@ -22,7 +27,7 @@ const MainLayout = ({ usersRole }: { usersRole: string }) => {
     return <Navigate to="/tenantdashboard" replace />;
 
   //disable the menu btn on specific routes
-  const disabledRoutes = ["/admin/manageTenant"];
+  const disabledRoutes = ["/admin/manageTenant", "/manageLease"];
   const showMenuButton = !disabledRoutes.includes(location.pathname);
 
   return (
@@ -37,7 +42,7 @@ const MainLayout = ({ usersRole }: { usersRole: string }) => {
 
       {/* Menu Button */}
 
-      {showMenuButton && (
+      {showMenuButton ? (
         <button
           onClick={toggleSidebar}
           className={`fixed pt-6 rounded-lg z-50 md:hidden ${
@@ -45,6 +50,13 @@ const MainLayout = ({ usersRole }: { usersRole: string }) => {
           } ml-4`}
         >
           <Menu size={32} />
+        </button>
+      ) : (
+        <button
+          className="rounded-full w-12 h-12 bg-[#d9d9d9] flex items-center justify-center absolute top-12 left-5 cursor-pointer"
+          onClick={goBack}
+        >
+          <ArrowLeft size={30} color="#000000" />
         </button>
       )}
