+++ frontend/src/components/LeaseCountdown.tsx
@@ -3,7 +3,7 @@ export default function LeaseCountdown() {
     <div className="w-full bg-black font-heading rounded-2xl p-5 md:w-1/2">
       <div className="flex items-center justify-between mb-4">
         <h1 className="text-white">Lease Countdown</h1>
-        <button className="bg-white text-black rounded-full px-4 py-2 font-semibold text-sm">
+        <button className="bg-white text-black rounded-full px-4 py-2 font-semibold text-sm cursor-pointer">
           Send Notification
         </button>
       </div>
