+++ frontend/src/data/issues.json
@@ -1,16 +1,16 @@
 [
-    {
-      "id": 1,
-      "date": "1/21/2025",
-      "title": "Toilet Maintainance",
-      "isNew": true,
-      "disabled": false
-    },
-    {
-      "id": 2,
-      "date": "12/21/2025",
-      "title": "Noise Complaint",
-      "isNew": false,
-      "disabled": true
-    }
-  ]
+  {
+    "id": 1,
+    "date": "1/21/2025",
+    "title": "Toilet Maintainance",
+    "isNew": true,
+    "disabled": false
+  },
+  {
+    "id": 2,
+    "date": "12/21/2025",
+    "title": "Noise Complaint",
+    "isNew": false,
+    "disabled": true
+  }
+]
