+++ frontend/src/pages/AdminManageTenant.tsx
@@ -1,17 +1,9 @@
-import { ArrowLeft } from "lucide-react";
-import { useNavigate } from "react-router-dom";
 import GuestProfileIcon from "../components/guests/GuestProfileIcon";
 import LeaseCountdown from "../components/LeaseCountdown";
 import { PackageCard } from "../tenantDashboard/components";
 import IssuesList from "../components/issues/IssueList";
 import CurrentGuestTable from "../components/guests/CurrentGuestTable";
 export default function AdminManageTenant() {
-  const navigate = useNavigate();
-
-  const goBack = () => {
-    navigate(-1);
-  };
-
   //TODO: package card component needs to have the add package btn if user is admin
   //- latest issues needs to pull data from user
 
@@ -19,12 +11,6 @@ export default function AdminManageTenant() {
 
   return (
     <div className="px-4 mt-14 font-heading min-h-screen space-y-6">
-      <button
-        className="rounded-full w-12 h-12 bg-[#d9d9d9] flex items-center justify-center absolute top-12 left-5"
-        onClick={goBack}
-      >
-        <ArrowLeft size={30} color="#000000" />
-      </button>
       <h1 className="text-center text-2xl font-heading md:text-left md:text-3xl">
         Manage Tenant
       </h1>
