+++ frontend/vite.config.ts
@@ -1,19 +1,17 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import tailwindcss from '@tailwindcss/vite'
-import fs from 'fs';
-import path from 'path';
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import tailwindcss from "@tailwindcss/vite";
+import fs from "fs";
+import path from "path";
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react(),
-  tailwindcss(),
-  ],
+  plugins: [react(), tailwindcss()],
   server: {
     https: {
-      key: fs.readFileSync(path.resolve(__dirname, './localhost-key.pem')),
-      cert: fs.readFileSync(path.resolve(__dirname, './localhost.pem')),
+      key: fs.readFileSync(path.resolve(__dirname, "./localhost-key.pem")),
+      cert: fs.readFileSync(path.resolve(__dirname, "./localhost.pem")),
     },
-    host: '0.0.0.0',
+    host: "0.0.0.0",
   },
-})
+});
